Add from_folder test for vision (#1003)
* add from_folder test for vision

* update test to mimic a more realistic scenario

* update tests with the updated MNIST_TINY dataset

* remove device dependency from tests

* Revert "remove device dependency from tests"

This reverts commit 87aeeeaa648db544ab028ce965a9cfaa81f04896.

* rename old test_vision to test_vision_cpu for CPU specific tests - and introduce a new test_vision for the rest of the stuff
