+++ tests/test_vision_data.py
@@ -1,0 +1,14 @@
+import pytest
+from fastai import *
+from fastai.vision import *
+
+@pytest.fixture(scope="module")
+def path():
+    path = untar_data(URLs.MNIST_TINY)
+    return path
+
+def test_from_folder(path):
+    for valid_pct in [None, 0.9]:
+        data = ImageDataBunch.from_folder(path, test='test', valid_pct=valid_pct)
+        assert len(data.classes) == 2
+        assert set(data.classes) == set(['3', '7'])
